# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для конструктора email-рассылок",
        "title": "Email Campaign API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
        "/api/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Логин",
                "parameters": [
                    {
                        "description": "Логин данные",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.LoginResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Регистрация пользователя",
                "parameters": [
                    {
                        "description": "Данные регистрации",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RegisterResponse"
                        }
                    }
                }
            }
        },
        "/api/campaigns": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "campaigns"
                ],
                "summary": "Получить все кампании",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetAllCampaignsResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "campaigns"
                ],
                "summary": "Создать кампанию",
                "parameters": [
                    {
                        "description": "Данные кампании",
                        "name": "campaign",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateCampaignRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateCampaignResponse"
                        }
                    }
                }
            }
        },
...
        "/api/campaigns/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "campaigns"
                ],
                "summary": "Получить кампанию по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID кампании",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CampaignResponse"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "campaigns"
                ],
                "summary": "Обновить статус кампании",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID кампании",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новый статус кампании",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateCampaignRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CampaignResponse"
                        }
                    }
                }
            }
        },
        "/api/campaigns/{id}/archive": {
            "patch": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "campaigns"
                ],
                "summary": "Архивировать кампанию",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID кампании",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ArchiveCampaignResponse"
                        }
                    }
                }
            }
        },
        "/api/campaigns/{id}/restore": {
            "patch": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "campaigns"
                ],
                "summary": "Восстановить кампанию из архива",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID кампании",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RestoreCampaignResponse"
                        }
                    }
                }
            }
        },
        "/api/clients": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clients"
                ],
                "summary": "Получить всех клиентов",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetAllClientsResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clients"
                ],
                "summary": "Создать клиента",
                "parameters": [
                    {
                        "description": "Данные клиента",
                        "name": "client",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateClientRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateClientResponse"
                        }
                    }
                }
            }
        },
...
        "/api/clients/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clients"
                ],
                "summary": "Получить клиента по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID клиента",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ClientResponse"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clients"
                ],
                "summary": "Обновить клиента",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID клиента",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные клиента",
                        "name": "client",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateClientRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateClientResponse"
                        }
                    }
                }
            }
        },
        "/api/clients/{id}/archive": {
            "patch": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clients"
                ],
                "summary": "Архивировать клиента",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID клиента",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ArchiveClientResponse"
                        }
                    }
                }
            }
        },
        "/api/clients/{id}/restore": {
            "patch": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clients"
                ],
                "summary": "Восстановить клиента из архива",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID клиента",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RestoreClientResponse"
                        }
                    }
                }
            }
        },
        "/api/templates": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "Получить все не архивированные шаблоны",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetAllTemplatesResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "Создать шаблон",
                "parameters": [
                    {
                        "description": "Данные шаблона",
                        "name": "template",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateTemplateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateTemplateResponse"
                        }
                    }
                }
            }
        },
...
        "/api/templates/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "Получить шаблон по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID шаблона",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetTemplateByIDResponse"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "Обновить шаблон",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID шаблона",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные шаблона",
                        "name": "template",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateTemplateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateTemplateResponse"
                        }
                    }
                }
            }
        },
        "/api/templates/{id}/archive": {
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "Архивировать шаблон",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID шаблона",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ArchiveTemplateResponse"
                        }
                    }
                }
            }
        },
        "/api/templates/{id}/restore": {
            "patch": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "Восстановить шаблон из архива",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID шаблона",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RestoreTemplateResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.ArchiveCampaignResponse": {
            "type": "object",
            "properties": {
                "campaign": {
                    "$ref": "#/definitions/dto.CampaignResponse"
                },
                "message": {
                    "type": "string",
                    "example": "кампания архивирована"
                }
            }
        },
        "dto.ArchiveClientResponse": {
            "type": "object",
            "properties": {
                "client": {
                    "$ref": "#/definitions/dto.ClientResponse"
                },
                "message": {
                    "type": "string",
                    "example": "клиент архивирован"
                }
            }
        },
...
        "dto.CampaignResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2023-10-10T10:00:00Z"
                },
                "created_by": {
                    "type": "integer",
                    "example": 2
                },
                "deleted_at": {
                    "type": "string",
                    "example": "2024-06-01T12:00:00Z"
                },
                "deleted_by": {
                    "type": "integer",
                    "example": 2
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "is_deleted": {
                    "type": "boolean",
                    "example": false
                },
                "scheduled_at": {
                    "type": "string",
                    "example": "2023-10-10T10:00:00Z"
                },
                "segment": {
                    "type": "string",
                    "example": "sport"
                },
                "status": {
                    "type": "string",
                    "example": "scheduled"
                },
                "tpl_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "dto.ClientResponse": {
            "type": "object",
            "properties": {
                "consent_flag": {
                    "type": "boolean",
                    "example": true
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-06-01T12:00:00Z"
                },
                "deleted_at": {
                    "type": "string",
                    "example": "2024-06-01T12:00:00Z"
                },
                "deleted_by": {
                    "type": "integer",
                    "example": 2
                },
                "email": {
                    "type": "string",
                    "example": "mirea@mail.ru"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "is_deleted": {
                    "type": "boolean",
                    "example": false
                },
                "segment": {
                    "type": "string",
                    "example": "sport"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-11-05T12:00:00Z"
                }
            }
        },
        "dto.CreateCampaignRequest": {
            "type": "object",
            "required": [
                "scheduled_at",
                "segment",
                "tpl_id"
            ],
            "properties": {
                "scheduled_at": {
                    "type": "string",
                    "example": "2023-10-10T10:00:00Z"
                },
                "segment": {
                    "type": "string",
                    "example": "sport"
                },
                "tpl_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "dto.CreateCampaignResponse": {
            "type": "object",
            "properties": {
                "campaign": {
                    "$ref": "#/definitions/dto.CampaignResponse"
                },
                "message": {
                    "type": "string",
                    "example": "кампания создана"
                }
            }
        },
        "dto.CreateClientRequest": {
            "type": "object",
            "required": [
                "email",
                "segment"
            ],
            "properties": {
                "consent_flag": {
                    "type": "boolean",
                    "example": true
                },
                "email": {
                    "type": "string",
                    "example": "mirea@mail.ru"
                },
                "segment": {
                    "type": "string",
                    "example": "sport"
                }
            }
        },
        "dto.CreateClientResponse": {
            "type": "object",
            "properties": {
                "client": {
                    "$ref": "#/definitions/dto.ClientResponse"
                },
                "message": {
                    "type": "string",
                    "example": "клиент создан"
                }
            }
        },
...
        "dto.CreateTemplateResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "шаблон создан"
                },
                "template": {
                    "$ref": "#/definitions/dto.TemplateResponse"
                }
            }
        },
        "dto.GetAllCampaignsResponse": {
            "type": "object",
            "properties": {
                "campaigns": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CampaignResponse"
                    }
                }
            }
        },
        "dto.GetAllClientsResponse": {
            "type": "object",
            "properties": {
                "clients": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ClientResponse"
                    }
                }
            }
        },
        "dto.GetAllTemplatesResponse": {
            "type": "object",
            "properties": {
                "templates": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.TemplateResponse"
                    }
                }
            }
        },
...
        "dto.LoginRequest": {
            "type": "object",
            "required": [
                "login",
                "password"
            ],
            "properties": {
                "login": {
                    "type": "string",
                    "example": "user123"
                },
                "password": {
                    "type": "string",
                    "example": "pass123"
                }
            }
        },
        "dto.LoginResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "успешный вход"
                },
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
            }
        },
        "dto.RegisterRequest": {
            "type": "object",
            "required": [
                "login",
                "password",
                "role"
            ],
            "properties": {
                "login": {
                    "type": "string",
                    "example": "user123"
                },
                "password": {
                    "type": "string",
                    "example": "pass123"
                },
                "role": {
                    "type": "string",
                    "enum": [
                        "marketer",
                        "analyst",
                        "admin"
                    ],
                    "example": "marketer"
                }
            }
        },
        "dto.RegisterResponse": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string",
                    "example": "user123"
                },
                "message": {
                    "type": "string",
                    "example": "пользователь создан"
                },
                "role": {
                    "type": "string",
                    "example": "marketer"
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "dto.RestoreCampaignResponse": {
            "type": "object",
            "properties": {
                "campaign": {
                    "$ref": "#/definitions/dto.CampaignResponse"
                },
                "message": {
                    "type": "string",
                    "example": "кампания восстановлена"
                }
            }
        },
        "dto.RestoreClientResponse": {
            "type": "object",
            "properties": {
                "client": {
                    "$ref": "#/definitions/dto.ClientResponse"
                },
                "message": {
                    "type": "string",
                    "example": "клиент восстановлен"
                }
            }
        },
...
        "dto.UpdateCampaignRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string",
                    "example": "scheduled"
                }
            }
        },
        "dto.UpdateClientRequest": {
            "type": "object",
            "properties": {
                "consent_flag": {
                    "type": "boolean",
                    "example": true
                },
                "email": {
                    "type": "string",
                    "example": "mirea@mail.ru"
                },
                "segment": {
                    "type": "string",
                    "example": "sport"
                }
            }
        },
        "dto.UpdateClientResponse": {
            "type": "object",
            "properties": {
                "client": {
                    "$ref": "#/definitions/dto.ClientResponse"
                },
                "message": {
                    "type": "string",
                    "example": "клиент обновлён"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "security": [
        {
            "BearerAuth": []
        }
    ]
}